requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
boto3==1.34.122
botocore==1.34.122
certifi==2023.7.22
charset-normalizer==3.2.0
dnspython==2.5.0
frozenlist==1.4.0
idna==3.4
jmespath==1.0.1
multidict==6.0.4
pymongo==3.11.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
requests==2.31.0
s3transfer==0.10.1
six==1.16.0
urllib3==2.0.4
yarl==1.9.2
